general transformation of plugins in a more generic way, documented in help of runnable and README.

+++ src/plugins/performance_plugin.h
@@ -74,6 +74,6 @@ int monitor_eventset_reset(void * monitor_eventset);
  * @param values, the array of values to update.
  * @return 0 on success to read events. -1 if read error occured.
  */  
-int monitor_eventset_read(void * monitor_eventset, double * values);
+int monitor_eventset_read(void * monitor_eventset, long long * values);
 
 #endif
